[Proposal] Add a lightweight proposal process (#49)
[Proposal] Add a lightweight proposal process

### Motivation

To ensure that we include all the important considerations for every non-trivial API change, and solicit feedback in a consistent way, let's introduce a lightweight proposal process based on Swift evolution.

### Modifications

Added a template and a Proposals page that lists them all, plus linked to it from the relevant doc pages.

### Result

Contributors will be able to just duplicate the template and fill in their proposal.

### Test Plan

Previewed docs locally and verified the navigation works.


Reviewed by: simonjbeaumont

Builds:
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-generator/pull/49