docs: Emphasize target setting in Xcode tutorial (#58)
### Motivation

In our Xcode tutorial we outline the steps to enable Swift OpenAPI
Generator on a separate framework. When you come to add the package
dependencies on the runtime and transport libraries the default target
Xcode offers to add them to is not this new framework. Failing to add
them to the framework target causes linker issues.

This is implied in our guide but it's pretty subtle and a user tripped
up over this.

### Modifications

- Emphasize which target the libraries need to be added to in the Xcode
tutorial.

### Result

Hopefully clearer step in the tutorial. 

### Test Plan

Ran through the steps manually and confirmed this generates a building
app.

### Resolves

Resolves #57.

Signed-off-by: Si Beaumont <beaumont@apple.com>