Support empty string OpenAPI identifiers (#59)
### Motivation

The OpenAPI specification allows the use of the empty string in some
places, for example in the following enum, which we've seen in some
real-world OpenAPI documents:

```yaml
type: string
enum:
  - ""
  - foo
  - bar
```

We have code that converts an OpenAPI identifier to a suitable string
that can be used as a Swift identifier. It currently returns an empty
string when given an empty string as input, which is not a valid Swift
identifier. We already have some logic there to avoid clashing with some
Swift keywords, so this can be extended to handle the empty string.

### Modifications

- Map the empty string OpenAPI identifier to "_empty"

### Result

We can generate compiling code when presented with empty string OpenAPI
identifiers.

### Test Plan

- Added unit test.
- Extended reference test.

### Resolves

- Resolves #33.
- Fixes #61.

---------

Signed-off-by: Si Beaumont <beaumont@apple.com>