docs: Update proposal process (#91)
### Motivation

The proposal process documentation currently suggests including the
implementation along with the proposal. Proposals make sense for
features that require up-front design or have complex or controversial
API implications. Changes that do not require up-front design are better
served with a pull-request for the change (and potentially additional
non-proposal documentation).

### Modifications

Adjust the wording of the proposal process to indicate that prototypes
should be on a separate branch for reference.

### Result

Proposals are made independently of the change.

### Test Plan

n/a.