Add finer-grain reference tests for translation phase (#103)
### Motivation

We have a reference test, which we've been making use of to validate
end-to-end generation from OpenAPI document matches handwritten Swift
files. This started out life as the canonical Petstore example from the
OpenAPI specification but has since been augmented in various ways to
test out some edge cases. This approach isn't scaling too well as we
have an all-or-nothing test.

We also don't have great test coverage over the translation phase of the
generator pipeline, which translates between the parsed OpenAPI document
to an intermediate representation.

### Modifications

Add test harnesses for finer grain reference tests that exercise smaller
sections of the translation phase. It should allow us to test a small
OpenAPI document renders to _equivalent_ Swift source (stripping
comments and reformatting).

### Result

Hopefully a more modular and maintainable testing strategy as we try to
support more of the OpenAPI specification.

### Test Plan

This PR _is_ tests.

---------

Signed-off-by: Si Beaumont <beaumont@apple.com>