[Tests] Add an integration test product for easier iteration (#121)
[Tests] Add an integration test product for easier iteration

### Motivation

Some IDEs don't allow you to trigger building of targets, only products.

### Modifications

Added an umbrella library product to make it possible to build locally in an IDE, not just using `swift build`.

### Result

Now the product shows up in IDEs, allowing developers to iterate on the code and rebuild.

### Test Plan

Tested that it shows up locally.


Reviewed by: simonjbeaumont

Builds:
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (docc test) - Build finished. 
     ✔︎ pull request validation (integration test) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-generator/pull/121