Move ContentType.identifier logic into the translator (#141)
Move ContentType.identifier logic into the translator

### Motivation

In preparation for changing the mapping of content type names onto enum case names, we need to have the `ContentType.identifier` logic on a translator, which has access to the feature flags, so that we can later conditionalize the logic.

### Modifications

Takes the first step - moves the existing logic onto the translator.

### Result

NFC

### Test Plan

Added unit tests both for the existing logic, and also for the stubbed out logic behind the feature flag.


Reviewed by: simonjbeaumont

Builds:
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (docc test) - Build finished. 
     ✔︎ pull request validation (integration test) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-generator/pull/141