Deprecation warnings not emitted when calling operations that are deprecated (#153)
Deprecation warnings not emitted when calling operations that are deprecated

### Motivation

Fixes #149.

Turns out that when a function is deprecated in a protocol, then calling the function on a concrete type that conforms to that protocol does _not_ emit a warning unless the function _implementation_ is also annotated as deprecated.

### Modifications

This PR emits deprecation annotations for the client and server implementations of `APIProtocol`'s functions.

### Result

Warnings are now correctly emitted regardless of whether the adopter calls methods on `let client: MyClient` where `MyClient: APIProtocol`, or on `let client: any APIProtocol`, leading to better communication between the author of the OpenAPI document and the adopter of the generated Swift code about the API evolution and future removal of operations.

### Test Plan

Adapted integration tests.


Reviewed by: simonjbeaumont

Builds:
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (docc test) - Build finished. 
     ✔︎ pull request validation (integration test) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-generator/pull/153