Add a proposal state for when merged behind a feature flag (#161)
Add a proposal state for when merged behind a feature flag

### Motivation

When introducing breaking changes, it can be convenient to hide the new behavior behind a feature flag. However, up until now, proposals didn't have a state that represents "feature landed but is not enabled by default yet".

### Modifications

This PR introduces a new state for proposals for this.

Also, added missing documentation for the `featureFlags` key in the config file.

### Result

Now proposals can be moved to the new state before their feature is enabled unconditionally.

### Test Plan

N/A


Reviewed by: simonjbeaumont

Builds:
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (docc test) - Build finished. 
     ✔︎ pull request validation (integration test) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-generator/pull/161