[Proposal] SOAR-0001 Improved OpenAPI -> Swift name mapping (#95)
### Motivation

Encoding a string would help distinguish types properly. For eg. both a
b and a_b would be rendered as a_b. This change would make it render
a_space_b and a_b respectively. Adding a proposal as its not a trivial
change, as recommended by the contribution guidelines.

### Modifications

Added proposal document - `SOAR-0001.md`

An initial draft of the proposed implementation is at
https://github.com/apple/swift-openapi-generator/pull/89. Any
suggestions or improvements for the proposal are welcome.