Forward schema warnings to diagnostics (#174)
Forward schema warnings to diagnostics

### Motivation

OpenAPIKit emits warnings about parsed `JSONSchema` objects that could be helpful to adopters when debugging issues. Today, we are ignoring them.

### Modifications

This PR forwards them into the diagnostics collector as warnings.

### Result

When a schema is malformed in a way that OpenAPIKit still manages to parse it, but has issues, adopters will see these issues emitted as all other diagnostics we emit.

### Test Plan

Added a unit test.


Reviewed by: gjcairo

Builds:
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (docc test) - Build finished. 
     ✔︎ pull request validation (integration test) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-generator/pull/174