[Docs] Clarify what constitutes a breaking change of generated code (#200)
[Docs] Clarify what constitutes a breaking change of generated code

### Motivation

As new features are introduced to the generator, it is limiting to not be able to add new defaulted properties to generated structs without it being an API-breaking change.

### Modifications

Clarify that adding a defaulted property to a generated struct is not considered a breaking change in the Swift OpenAPI Generator project, and adopters should avoid capturing initializers of generated structs for that reason.

### Result

With this clarification, we can add new defaulted properties without breaking API.

### Test Plan

N/A


Reviewed by: glbrntt

Builds:
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (docc test) - Build finished. 
     ✔︎ pull request validation (integration test) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-generator/pull/200