[Generator] Stop generating an undocumented case for enums/oneOfs (#205)
### Motivation

Fixes https://github.com/apple/swift-openapi-generator/issues/204.

Depends on https://github.com/apple/swift-openapi-runtime/pull/40.

### Modifications

This introduces the feature flag `closedEnumsAndOneOfs` which stops
generating the `undocumented` case for enums and oneOfs (no other places
are affected, such as responses, and so on.). The default behavior in
0.1.x stays the same, but in 0.2.0 we'll stop generating `undocumented`
cases unconditionally.

This fixes our divergence from the OpenAPI specification, as we treated
enums and oneOfs as open, but they're closed by default.

I also added docs on this pattern to make it easier for adopters to know
how to create open enums and oneOfs on purpose.

### Result

We get closer to being spec-compliant.

### Test Plan

Created new snippet tests to verify the new functionality, will adapt
file-based tests when we're preparing 0.2.0.