Allow refs to oneOf/anyOf inside a oneOf with a discriminator (#224)
Allow refs to oneOf/anyOf inside a oneOf with a discriminator

### Motivation

Fixes #222.

We had a bug, where we considered oneOf/anyOf nested within a oneOf with a discriminator as not supported.

We already allowed allOf in there, but it was an oversight to disallow oneOf/anyOf there.

### Modifications

Update the isSupported logic to allow oneOf and anyOf within such oneOfs with a discriminator. Note that the nested schemas _within_ the oneOf/anyOf still must be objectish, as oneOf + discriminator always represent a JSON object.

### Result

OpenAPI docs with this nesting won't skip such schemas anymore.

### Test Plan

Updated the unit test.


Reviewed by: glbrntt

Builds:
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (docc test) - Build finished. 
     ✔︎ pull request validation (integration test) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-generator/pull/224