Allow non-unique values in oneOf.discriminator.mapping (#225)
Allow non-unique values in oneOf.discriminator.mapping

### Motivation

Fixes #223, see the issue for details.

### Modifications

Changed the logic for deriving the enum cases to take both the mapping and the list of subschemas into account. Documented it more in code.

### Result

Now adopter can actually have different keys pointing to the same types in their mapping - originally reported by an adopter, so it's a real issue.

### Test Plan

Adapted tests.

Reviewed by: glbrntt

Builds:
     ✔︎ pull request validation (5.8) - Build finished. 
     ✔︎ pull request validation (5.9) - Build finished. 
     ✔︎ pull request validation (docc test) - Build finished. 
     ✔︎ pull request validation (integration test) - Build finished. 
     ✔︎ pull request validation (nightly) - Build finished. 
     ✔︎ pull request validation (soundness) - Build finished. 

https://github.com/apple/swift-openapi-generator/pull/225