Prepare 0.2.0 release (API breaking changes) (#218)
Prepare 0.2.0 release (API breaking changes)

### Motivation

It's time to release an API-breaking version and delete all the deprecated code.

Requires that we first release 0.2.0 of the runtime library.

### Modifications

Deleted all the deprecated code, removed all the feature flags and 

This is a breaking change, so the integration test will fail until we land this and then the corresponding generator changes.

### Result

Removed a bunch of code!

### Test Plan

Let's let CI check everything except for the integration test.


Reviewed by: gjcairo

Builds:
     ✔︎ pull request validation (soundness) - Build finished. 
     ✖︎ pull request validation (5.8) - Build finished. 
     ✖︎ pull request validation (5.9) - Build finished. 
     ✖︎ pull request validation (docc test) - Build finished. 
     ✖︎ pull request validation (integration test) - Build finished. 
     ✖︎ pull request validation (nightly) - Build finished. 

https://github.com/apple/swift-openapi-generator/pull/218